requirements.txt
asgiref==3.4.1
coverage==6.0.1
Django==3.2.5
django-widget-tweaks==1.4.8
Faker==8.10.3
libgravatar==1.0.0
numpy==1.21.4
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
django-location-field==2.1.0
django-countries==7.2.1
